# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Voting System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  /** 12-digit Aadhaar number */
  aadhaarNumber: string;
  /** 128-dimension face descriptor array */
  faceDescriptor?: number[] | null;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  voterId: string;
  name: string;
  email: string;
  token: string;
  hasVoted: boolean;
  aadhaarVerified: boolean;
}

export interface VoterProfile {
  voterId: string;
  name: string;
  email: string;
  hasVoted: boolean;
  aadhaarVerified: boolean;
  hasFaceData: boolean;
  registeredAt: string;
}

export interface VerifyAadhaarRequest {
  voterId: string;
  aadhaarNumber: string;
}

export interface AadhaarVerifyResponse {
  verified: boolean;
  message: string;
}

export interface StoreFaceRequest {
  voterId: string;
  faceDescriptor: number[];
}

export interface Candidate {
  id: number;
  name: string;
  party: string;
  description: string;
  photoUrl?: string | null;
  voteCount: number;
  symbol?: string | null;
}

export interface CreateCandidateRequest {
  name: string;
  party: string;
  description: string;
  photoUrl?: string | null;
  symbol?: string | null;
}

export interface CastVoteRequest {
  voterId: string;
  candidateId: number;
  faceVerified: boolean;
  aadhaarVerified: boolean;
}

export interface VoteResponse {
  success: boolean;
  blockHash: string;
  blockIndex: number;
  message: string;
}

export interface VotingStatus {
  hasVoted: boolean;
  votedAt?: string | null;
  candidateId?: number | null;
}

export interface CandidateResult {
  candidateId: number;
  name: string;
  party: string;
  symbol?: string | null;
  voteCount: number;
  percentage: number;
}

export interface ElectionResults {
  isHidden: boolean;
  totalVotes: number;
  results: CandidateResult[];
}

export interface Block {
  index: number;
  voterId: string;
  candidateId: number;
  timestamp: string;
  previousHash: string;
  hash: string;
  nonce: number;
}

export interface BlockchainResponse {
  blocks: Block[];
  isValid: boolean;
  totalBlocks: number;
}

export interface BlockchainValidation {
  isValid: boolean;
  message: string;
  brokenAt?: number | null;
}

export interface ElectionStatus {
  isActive: boolean;
  startedAt?: string | null;
  endedAt?: string | null;
  resultsHidden: boolean;
}

export interface UpdateElectionRequest {
  isActive: boolean;
}

export interface AdminStats {
  totalVoters: number;
  totalVotesCast: number;
  fraudAttempts: number;
  activeElection: boolean;
  verificationRate: number;
}

export interface ResetRequest {
  adminPassword: string;
}

export interface FraudLog {
  id: number;
  voterId?: string | null;
  attemptType: string;
  details: string;
  ipAddress?: string | null;
  createdAt: string;
}

export interface LogFraudRequest {
  voterId?: string | null;
  attemptType: string;
  details: string;
}

export interface ToggleResultsRequest {
  hidden: boolean;
}