# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Digital Voting System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerVoter
      tags: [auth]
      summary: Register a new voter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Voter registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Voter already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginVoter
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-aadhaar:
    post:
      operationId: verifyAadhaar
      tags: [auth]
      summary: Verify Aadhaar number for logged-in voter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyAadhaarRequest"
      responses:
        "200":
          description: Aadhaar verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AadhaarVerifyResponse"
        "401":
          description: Aadhaar mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/store-face:
    post:
      operationId: storeFaceData
      tags: [auth]
      summary: Store face descriptor for voter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreFaceRequest"
      responses:
        "200":
          description: Face data stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getCurrentVoter
      tags: [auth]
      summary: Get current voter info
      parameters:
        - in: header
          name: x-voter-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Voter info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoterProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateRequest"
      responses:
        "201":
          description: Candidate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    put:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateRequest"
      responses:
        "200":
          description: Candidate updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/election:
    get:
      operationId: getElectionStatus
      tags: [admin]
      summary: Get election status
      responses:
        "200":
          description: Election status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionStatus"
    put:
      operationId: updateElectionStatus
      tags: [admin]
      summary: Start or stop election
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateElectionRequest"
      responses:
        "200":
          description: Election status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionStatus"
...
  /admin/reset:
    post:
      operationId: resetElection
      tags: [admin]
      summary: Reset the entire election
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetRequest"
      responses:
        "200":
          description: Election reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/fraud-log:
    get:
      operationId: getFraudLog
      tags: [admin]
      summary: Get fraud attempt logs
      responses:
        "200":
          description: Fraud logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FraudLog"
    post:
      operationId: logFraudAttempt
      tags: [admin]
      summary: Log a fraud attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogFraudRequest"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/toggle-results:
    put:
      operationId: toggleResultsVisibility
      tags: [admin]
      summary: Toggle results visibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleResultsRequest"
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        aadhaarNumber:
          type: string
          description: 12-digit Aadhaar number
        faceDescriptor:
          type: array
          items:
            type: number
          nullable: true
          description: 128-dimension face descriptor array
      required:
        - name
        - email
        - password
        - aadhaarNumber
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        voterId:
          type: string
        name:
          type: string
        email:
          type: string
        token:
          type: string
        hasVoted:
          type: boolean
        aadhaarVerified:
          type: boolean
      required:
        - voterId
        - name
        - email
        - token
        - hasVoted
        - aadhaarVerified
    VoterProfile:
      type: object
      properties:
        voterId:
          type: string
        name:
          type: string
        email:
          type: string
        hasVoted:
          type: boolean
        aadhaarVerified:
          type: boolean
        hasFaceData:
          type: boolean
        registeredAt:
          type: string
      required:
        - voterId
        - name
        - email
        - hasVoted
        - aadhaarVerified
        - hasFaceData
        - registeredAt
    VerifyAadhaarRequest:
      type: object
      properties:
        voterId:
          type: string
        aadhaarNumber:
          type: string
      required:
        - voterId
        - aadhaarNumber
    AadhaarVerifyResponse:
      type: object
      properties:
        verified:
          type: boolean
        message:
          type: string
      required:
        - verified
        - message
    StoreFaceRequest:
      type: object
      properties:
        voterId:
          type: string
        faceDescriptor:
          type: array
          items:
            type: number
      required:
        - voterId
        - faceDescriptor
    Candidate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        party:
          type: string
        description:
          type: string
        photoUrl:
          type: string
          nullable: true
        voteCount:
          type: integer
        symbol:
          type: string
          nullable: true
      required:
        - id
        - name
        - party
        - description
        - voteCount
    CreateCandidateRequest:
      type: object
      properties:
        name:
          type: string
        party:
          type: string
        description:
          type: string
        photoUrl:
          type: string
          nullable: true
        symbol:
          type: string
          nullable: true
      required:
        - name
        - party
        - description
...
    ElectionStatus:
      type: object
      properties:
        isActive:
          type: boolean
        startedAt:
          type: string
          nullable: true
        endedAt:
          type: string
          nullable: true
        resultsHidden:
          type: boolean
      required:
        - isActive
        - resultsHidden
    UpdateElectionRequest:
      type: object
      properties:
        isActive:
          type: boolean
      required:
        - isActive
...
    ResetRequest:
      type: object
      properties:
        adminPassword:
          type: string
      required:
        - adminPassword
    FraudLog:
      type: object
      properties:
        id:
          type: integer
        voterId:
          type: string
          nullable: true
        attemptType:
          type: string
        details:
          type: string
        ipAddress:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - attemptType
        - details
        - createdAt
    LogFraudRequest:
      type: object
      properties:
        voterId:
          type: string
          nullable: true
        attemptType:
          type: string
        details:
          type: string
      required:
        - attemptType
        - details
    ToggleResultsRequest:
      type: object
      properties:
        hidden:
          type: boolean
      required:
        - hidden