# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Voting System API
 * OpenAPI spec version: 0.1.0
 */

export interface Candidate {
  id: number;
  name: string;
  party: string;
  description: string;
  photoUrl?: string | null;
  voteCount: number;
  symbol?: string | null;
}