# lib/api-zod/src/generated/types/storeFaceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Voting System API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreFaceRequest {
  voterId: string;
  faceDescriptor: number[];
}