# lib/api-zod/src/generated/types/verifyAadhaarRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Voting System API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyAadhaarRequest {
  voterId: string;
  aadhaarNumber: string;
}