# lib/api-zod/src/generated/types/voterProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital Voting System API
 * OpenAPI spec version: 0.1.0
 */

export interface VoterProfile {
  voterId: string;
  name: string;
  email: string;
  hasVoted: boolean;
  aadhaarVerified: boolean;
  hasFaceData: boolean;
  registeredAt: string;
}